refactor(viewer): extract shared batched helpers; drop dead overlay-subset path
Two duplicated helpers pulled into shared modules, and the dead
selection/preselection overlay-subset branch removed now that those
highlight by recoloring instances in place (batchedHighlight), never via
a subset.

- conwayVector.js: forEachVectorItem (Conway Vector / plain Array
  iteration) was byte-identical in flatMeshToBatchedModel and
  flatMeshToInstancedModel; both now import it.
- batchedModel.js: eachBatch (walk each BatchedMesh in a mesh-or-Group
  model) + isBatchedModel moved out of batchedHighlight; batchedSubset's
  subset walk and glbExport's modelHasBatchedMesh now route through
  eachBatch instead of re-implementing the traverse. batchedHighlight
  re-exports isBatchedModel so existing importers/tests are unchanged.
- batchedSubset: drop OVERLAY_CUSTOM_IDS / opts.raycastInvisible --
  isolation is the only caller that reaches this builder, so every subset
  stays pickable; test updated to assert that.
